+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cat48: Add ModeCCodeAndConfidenceIndicator
 - Cat48: Add HeightMeasuredBy3dRadar
 - Cat48: Add RadialDopplerSpeed
+- Cat48: Add ACASResolutionAdvisoryReport
 - Update Data Item docs
 
 ## [0.2.0] - 2020-08-21